Move components directory
