+++ app/src/main/java/com/github/leuludyha/ibdb/presentation/components/books/reading_list/ReadingList.kt
@@ -1,4 +1,4 @@
-package com.github.leuludyha.ibdb.presentation.components.reading_list
+package com.github.leuludyha.ibdb.presentation.components.books.reading_list
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
@@ -11,7 +11,7 @@ import androidx.navigation.NavHostController
 import com.github.leuludyha.domain.model.user.UserPreferences
 import com.github.leuludyha.domain.model.user.WorkPreference
 import com.github.leuludyha.ibdb.presentation.components.ItemList
-import com.github.leuludyha.ibdb.presentation.components.book_views.MiniBookView
+import com.github.leuludyha.ibdb.presentation.components.books.book_views.MiniBookView
 import com.github.leuludyha.ibdb.presentation.navigation.Screen
 
 @OptIn(ExperimentalMaterial3Api::class)
