+++ app/src/main/java/com/github/leuludyha/ibdb/presentation/screen/collection/CollectionScreen.kt
@@ -10,7 +10,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.github.leuludyha.ibdb.R
 import com.github.leuludyha.ibdb.presentation.components.authentication.DisplayIfAuthenticated
-import com.github.leuludyha.ibdb.presentation.components.reading_list.ReadingList
+import com.github.leuludyha.ibdb.presentation.components.books.reading_list.ReadingList
 
 @Composable
 fun CollectionScreen(
