+++ app/src/main/java/com/github/leuludyha/ibdb/presentation/screen/search/BookSearchScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.SideEffect
 import androidx.navigation.NavHostController
 import com.github.leuludyha.ibdb.presentation.Orientation
 import com.github.leuludyha.ibdb.presentation.components.PagingItemList
-import com.github.leuludyha.ibdb.presentation.components.book_views.MiniBookView
+import com.github.leuludyha.ibdb.presentation.components.books.book_views.MiniBookView
 import com.github.leuludyha.ibdb.presentation.components.search.BookSearch
 import com.github.leuludyha.ibdb.presentation.navigation.Screen
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
